+++ d10/ex06/srcs/ft_usage.c
@@ -6,13 +6,13 @@
 /*   By: xamartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/14 21:52:23 by xamartin          #+#    #+#             */
-/*   Updated: 2017/09/16 20:49:34 by xamartin         ###   ########.fr       */
+/*   Updated: 2017/09/18 16:40:19 by xamartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_header.h"
 
-long	ft_usage(long a, long b)
+int	ft_usage(int a, int b)
 {
 	(void)a;
 	(void)b;
