+++ d10/ex09/includes/ft_header.h
@@ -6,7 +6,7 @@
 /*   By: xamartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/14 21:43:26 by xamartin          #+#    #+#             */
-/*   Updated: 2017/09/15 22:56:23 by xamartin         ###   ########.fr       */
+/*   Updated: 2017/09/18 16:46:29 by xamartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 
 # include <unistd.h>
 
-int						ft_add(long a, long b);
-int						ft_sub(long a, long b);
-int						ft_mul(long a, long b);
-int						ft_div(long a, long b);
-int						ft_mod(long a, long b);
-int						ft_usage(long a, long b);
+int						ft_add(int a, int b);
+int						ft_sub(int a, int b);
+int						ft_mul(int a, int b);
+int						ft_div(int a, int b);
+int						ft_mod(int a, int b);
+int						ft_usage(int a, int b);
 int						ft_strcmp(char *s1, char *s2);
-long					ft_atoi(char *str);
+int						ft_atoi(char *str);
 void					ft_putnbr(int nb);
 void					ft_putchar(char c);
 void					zero(int i);
@@ -30,7 +30,7 @@ void					zero(int i);
 typedef struct			s_opp
 {
 	char				*str;
-	int					(*f)(long a, long b);
+	int					(*f)(int a, int b);
 }						t_opp;
 
 #endif
