+++ d10/ex09/srcs/calcul.c
@@ -6,33 +6,33 @@
 /*   By: xamartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/14 21:44:24 by xamartin          #+#    #+#             */
-/*   Updated: 2017/09/15 23:08:23 by xamartin         ###   ########.fr       */
+/*   Updated: 2017/09/19 16:45:37 by xamartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_header.h"
 
-int		ft_add(long a, long b)
+int		ft_add(int a, int b)
 {
-	return ((int)(a + b));
+	return (a + b);
 }
 
-int		ft_sub(long a, long b)
+int		ft_sub(int a, int b)
 {
-	return ((int)(a - b));
+	return (a - b);
 }
 
-int		ft_mul(long a, long b)
+int		ft_mul(int a, int b)
 {
-	return ((int)(a * b));
+	return (a * b);
 }
 
-int		ft_div(long a, long b)
+int		ft_div(int a, int b)
 {
-	return ((int)(a / b));
+	return (a / b);
 }
 
-int		ft_mod(long a, long b)
+int		ft_mod(int a, int b)
 {
-	return ((int)(a % b));
+	return (a % b);
 }
